refactor(cli): refactor bead & chain folder
