+++ packages/cli/core/compile/bead/ConfigBead.js
