+++ packages/cli/core/compile/bead/ScriptBead.js
