+++ packages/cli/core/compile/bead/TemplateBead.js
