+++ packages/cli/core/compile/bead/WepyBead.js
