+++ packages/cli/core/compile/bead/index.js
@@ -1,0 +1,6 @@
+exports.Bead = require('./Bead');
+exports.ConfigBead = require('./ConfigBead');
+exports.ScriptBead = require('./ScriptBead');
+exports.TemplateBead = require('./TemplateBead');
+exports.StyleBead = require('./StyleBead');
+exports.WepyBead = require('./WepyBead');
