+++ packages/cli/core/compile/chain/ComponentChain.js
