+++ packages/cli/core/compile/chain/WeappChain.js
