+++ packages/cli/core/compile/chain/WepyChain.js
