+++ packages/cli/core/compile/chain/index.js
@@ -1,0 +1,7 @@
+exports.Chain = require('./Chain');
+exports.AppChain = require('./AppChain');
+exports.ComponentChain = require('./ComponentChain');
+// exports.JsChain = require('./JsChain');
+exports.PageChain = require('./PageChain');
+exports.WeappChain = require('./WeappChain');
+exports.WepyChain = require('./WepyChain');
